[project @ Arch-1:robey@lag.net--2003-public%secsh--dev--1.0--patch-59] speed up parts of BufferedFile BufferedFile uses cStringIO for the write buffer now (i don't actually notice any speed difference so this might revert later) and the default buffer size has been upped from 1KB to 8KB.
when scanning for linefeeds (when writing to a line-buffered file), only scan
the newly-written bytes, since we know all the previously buffered data is
linefeed-free.  this was the #1 slowdown on the 1MB-file unit test.

also, limit the buffering on line-buffered files to whatever the default
buffer size is.  there's no reason to buffer 1MB waiting for a linefeed.